// src/app/interfaces/req-response.ts
// Generated by https://quicktype.io

export interface UsersResponse {
  page: number;
  per_page: number;
  total: number;
  total_pages: number;
  data: User[];
  support: Support;
}

export interface UserResponse {
  data: User;
  support: Support;
}

export interface User {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  avatar: string;
}

export interface Support {
  url: string;
  text: string;
}

export interface UsuarioResponse {
  usuario: Usuario;
  ok: boolean;
}

export interface Usuario {
  email: string;
  password: string;
}

export interface SalaResponse {
  host: string;
  sala: string;
  ok: boolean;
}


export interface DataSalaResponse {
  sala: string;
  data: string;
  ok: boolean;
}


export interface DataPaper {
  cells: Cell[];
}

export interface Cell {
  type: string;
  z: number;
  size?: Size;
  position?: Position;
  angle?: number;
  id: string;
  embeds?: string[];
  attrs: CellAttrs;
  source?: Source;
  target?: Target;
  parent?: string;
  labels?: LabelElement[];
}

export interface CellAttrs {
  label?: LabelTextClass;
}

export interface LabelTextClass {
  text: string;
}

export interface LabelElement {
  attrs: LabelAttrs;
}

export interface LabelAttrs {
  labelText: LabelTextClass;
}

export interface Position {
  x: number;
  y: number;
}

export interface Size {
  width: number;
  height: number;
}

export interface Source {
  id: string;
  anchor?: SourceAnchor;
}

export interface SourceAnchor {
  name: string;
  args: Args;
}

export interface Args {
  length: number;
}

export interface Target {
  x?: number;
  y?: number;
  id?: string;
  anchor?: TargetAnchor;
}

export interface TargetAnchor {
  name: string;
}


export interface GraphToCodigo {
  infoCodigo: string;
  ok: boolean;
}

export interface UnirseSalaXcodigo {
  sala: string;
  ok: boolean;
}


export interface ColaboradoresResponse {
  asistentes: Asistente[];
  sala:       string;
  ok:         boolean;
}

export interface Asistente {
  email: string;
}